requirements.txt
pygame-ce
noise
numpy